third_party/WebKit/LayoutTests/accessibility/editable-root.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div id="main" role="main">

  <input id="input" type="text" value="Line 1">

  <div id="contenteditable-textbox" role="textbox" contenteditable="true">
    <div id="contenteditable-inner" contenteditable>Line 1</div>
    <textarea id="contenteditable-line2" rows="1" cols="40">Line 2</textarea>
  </div>

  <div id="contenteditable-div" contenteditable>
    <p id="paragraph1">Line 1<br>
      Line 2</p>
    <p id="paragraph2">Line 3</p>
  </div>

</div>

<script>
  test(function() {
    var axInput = accessibilityController.accessibleElementById('input');
    assert_equals(axInput.editableRoot, axInput);
  }, 'The editable root of an input element is the input element itself.');

  test(function() {
    var axContentEditable = accessibilityController.accessibleElementById('contenteditable-textbox');
    assert_equals(axContentEditable.editableRoot, axContentEditable);
  }, 'The editable root of a content editable with a role of textbox is the content editable itself.');

  test(function() {
    var axInnerContentEditable = accessibilityController.accessibleElementById('contenteditable-inner');
    assert_equals(axInnerContentEditable.editableRoot, axInnerContentEditable);
  }, 'The editable root of a content editable inside another content editable is the inner content editable.');

  test(function() {
    var axTextarea = accessibilityController.accessibleElementById('contenteditable-line2');
    assert_equals(axTextarea.editableRoot, axTextarea);
  }, 'The editable root of a textarea embedded inside a content editable is the textarea itself.');

  test(function() {
    var axContentEditable = accessibilityController.accessibleElementById('contenteditable-div');
    var axParagraph1 = accessibilityController.accessibleElementById('paragraph1');
    var axParagraph2 = accessibilityController.accessibleElementById('paragraph2');
    assert_equals(axParagraph1.editableRoot, axContentEditable);
    assert_equals(axParagraph2.editableRoot, axContentEditable);
  }, 'The editable root of a paragraph within a content editable is the content editable.');

  if (window.testRunner)
    document.getElementById("main").style.display = "none";
</script>
